chatur_core/requirements.txt
# Core dependencies
numpy>=1.21.0
tensorflow>=2.8.0
beautifulsoup4>=4.11.0
scikit-learn>=1.1.0
sentence-transformers>=2.2.0
groq>=0.4.0

# PDF processing (at least one required)
PyPDF2>=3.0.0
pypdf>=3.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Telegram bot framework
python-telegram-bot>=20.0

# Additional implicit dependencies
torch>=1.12.0
transformers>=4.21.0
requests>=2.28.0
